[refactor] 폴더 경로 설정
