+++ frontend/Idle/src/components/Navbar/EstimatePrice.jsx
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import { styled } from "styled-components";
-import { carContext } from "../../../utils/context";
+import { carContext } from "../../utils/context";
 
 function EstimatePrice() {
   const { car } = useContext(carContext);
