+++ frontend/Idle/src/components/Navbar/Navbar.jsx
@@ -1,5 +1,5 @@
 import NavbarBox from "./NavbarBox.jsx";
-import { TYPE } from "../../../utils/constants.jsx";
+import { TYPE } from "../../utils/constants.jsx";
 import { styled } from "styled-components";
 import EstimatePrice from "./EstimatePrice.jsx";
