+++ frontend/Idle/src/components/common/MainLogo.jsx
@@ -1,7 +1,6 @@
-// import MainLogoImg from "../../assets/images/Hyundai.svg";
 import styled from "styled-components";
 import { ReactComponent as MainLogoImg } from "../../assets/images/hyundai.svg";
-import { HYUNDAI_MAIN } from "../../../utils/routes";
+import { HYUNDAI_MAIN } from "../../utils/routes";
 
 function MainLogo() {
   function logoClicked() {
