--- frontend/Idle/src/components/trim/Trim.jsx
@@ -1,5 +1,0 @@
-function Trim() {
-  return <div>Trim</div>;
-}
-
-export default Trim;
