+++ frontend/Idle/src/pages/colorPage.jsx
@@ -1,5 +1,11 @@
+import Navbar from "../components/Navbar/Navbar";
+
 function ColorPage() {
-  return <div>ColorPage</div>;
+  return (
+    <div>
+      <Navbar />
+    </div>
+  );
 }
 
 export default ColorPage;
