+++ frontend/Idle/src/pages/trimPage.jsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from "react";
 import { getTrimData } from "../utils/api";
 import TrimBoxContainer from "../components/TrimBoxContainer/TrimBoxContainer";
+import MainLogo from "../components/common/MainLogo";
 
 function TrimPage() {
   const [trimData, setTrimData] = useState(null);
@@ -11,7 +12,12 @@ function TrimPage() {
     });
   }, []);
 
-  return <div>{trimData ? <TrimBoxContainer {...trimData} /> : <p>Loading...</p>}</div>;
+  return (
+    <div>
+      {trimData ? <TrimBoxContainer {...trimData} /> : <p>Loading...</p>}
+      <MainLogo />
+    </div>
+  );
 }
 
 export default TrimPage;
